// src/rc-admin/types.ts
export type RecordToStringFunction = (record: any) => string;

export interface ResourceDefinition<OptionsType extends ResourceOptions = any> {
  readonly name: string;
  readonly options?: OptionsType;
  readonly hasList?: boolean;
  readonly hasEdit?: boolean;
  readonly hasShow?: boolean;
  readonly hasCreate?: boolean;
  readonly icon?: any;
  readonly recordRepresentation?:
    | ReactElement
    | RecordToStringFunction
    | string;
}
// ...
export interface ResourceOptions {
  label?: string;
  [key: string]: any;
}

export interface ResourceProps {
  intent?: "route" | "registration";
  name: string;
  list?: ComponentType<any> | ReactElement;
  create?: ComponentType<any> | ReactElement;
  edit?: ComponentType<any> | ReactElement;
  show?: ComponentType<any> | ReactElement;
  hasCreate?: boolean;
  hasEdit?: boolean;
  hasShow?: boolean;
  icon?: ComponentType<any>;
  recordRepresentation?: ReactElement | RecordToStringFunction | string;
  options?: ResourceOptions;
  children?: ReactNode;
}

export type Exporter = (
  data: any,
  fetchRelatedRecords: (
    data: any,
    field: string,
    resource: string
  ) => Promise<any>,
  dataProvider: DataProvider,
  resource?: string
) => void | Promise<void>;

export type SetOnSave = (
  onSave?: (values: object, redirect: any) => void
) => void;

export type FormFunctions = {
  setOnSave?: SetOnSave;
};

export type AdminRouterStatus = "loading" | "empty" | "ready";
export type ResourceWithRegisterFunction = {
  registerResource: (
    props: ResourceProps,
    permissions: any
  ) => ResourceDefinition;
};
export type RoutesAndResources = {
  customRoutesWithLayout: ReactElement<CustomRoutesProps>[];
  customRoutesWithoutLayout: ReactElement<CustomRoutesProps>[];
  resources: (ReactElement<ResourceProps> & ResourceWithRegisterFunction)[];
};